pyproject.toml
[tool.poetry]
name = "yarrow"
version = "0.1.0"
description = ""
authors = ["Dmitriy Shikhalev <dmitriy.shikhalev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pika = "^1.3.2"
pydantic = "^2.3.0"
pydantic-settings = "^2.0.3"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
pytest-mypy = "^0.10.3"
pytest-pylint = "^0.19.0"
pytest-aiohttp = "^1.0.4"
pytest-mock = "^3.10.0"
pytest-cov = "^4.0.0"
freezegun = "^1.2.2"
pika-stubs = "^0.1.3"
types-pyyaml = "^6.0.12.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
yarrow = "yarrow.main:serve"
